requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
Pillow>=8.3.0
instagrapi>=1.16.0
requests>=2.25.0
schedule>=1.1.0
python-dateutil>=2.8.0
feedparser>=6.0.0
beautifulsoup4>=4.9.0
lxml>=4.6.0